requirements.txt
scikit-learn==0.21.2
sphinx
numpydoc
pytest
flake8